refactor xapers.parser[s]
This introduces a dictionary
xapers.parsers.DOCUMENT_PARSERS that contains
a mapping from mimetypes to doc2txt parsers.
To use this, xapers.parser.parse_file
is also moved to xapers.parsers.__init__.

Also, the parts that use parse_file are adjusted accordingly.
